+++ test/WebJobs.Script.Tests/Workers/Rpc/WorkerConcurrencyManagerTest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -40,7 +40,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Workers
             _loggerFactory.AddProvider(_loggerProvider);
             _testEnvironment = new TestEnvironment();
             _testEnvironment.SetEnvironmentVariable(RpcWorkerConstants.FunctionsWorkerDynamicConcurrencyEnabled, "true");
-            _testEnvironment.SetEnvironmentVariable(RpcWorkerConstants.FunctionWorkerRuntimeSettingName, RpcWorkerConstants.PythonLanguageWorkerName);
             _testEnvironment.SetEnvironmentVariable(EnvironmentSettingNames.FunctionWorkerRuntime, RpcWorkerConstants.NodeLanguageWorkerName);
 
             var optionsMonitor = new Mock<IOptionsMonitor<FunctionsHostingConfigOptions>>();